Add tracked community info matching
This allows tracking (& banning) communities based on their status/info data. It's not the most thorough solution, but necessary for servers that are on inconsistent residential IPs and such.

Matching is just done by using PostgreSQL jsonb_path_exist() with a configurable path.

Also made a test page because I don't trust anybody to be able to get this down first-try.