Make can_skip_build query parameter have value.
The hub sends a can_skip_build parameter to indicate ACZ doesn't need to be built. Well I sent it without value originally because I figured that wasn't necessary to indicate boolean state.

Wow I made the critical mistake of trusting a 150 upvote Stack Overflow answer saying that's valid.

According to Stack Overflow this should be "totally valid" but what they fail to mention is that "totally valid" means "there is no spec for or against" which means that "does your server-side framework sanely support it" is not really guaranteed. Making a Stack Overflow answer like this is just irresponsible Jesus Christ.