Account logging system rewrite.
Now stores actor in separate database column to make it easier to access, it's too common.

Store acting IP address. Intending to implement a cron to remove this automatically after 2 weeks.

Rewrite code so it can be shared with hub audit logging and is less of a chore to write.