Make the default page the manage page
The current index page is kinda useless, theres no useful information on it and most people are just logging in for the account management options. There is no point to the index page other then create confussion on how to get to the account settings page (It is not very obvious you have to click on the "Hello XYZ" link) right now and is much better to just redirect to the Management page directly.

If you are not logged in you will be sent to the login screen, where theres 2 different places to click to signup for a new user. Hard to miss.